use float 64

+++ pyvista/core/grid.py
@@ -193,7 +193,7 @@ class RectilinearGrid(_vtk.vtkRectilinearGrid, Grid):
 
         # weird side effects if x is not floating point
         if not isinstance(x.dtype, np.floating):
-            x = x.astype(np.float32)
+            x = x.astype(np.float64)
         self.SetXCoordinates(helpers.convert_array(x.ravel()))
         if y is not None:
             if check_duplicates:
